Split out getSerializedPostsQuery
