+++ client/state/posts/utils/get-serialized-posts-query.js
@@ -1,0 +1,22 @@
+/**
+ * Internal dependencies
+ */
+import { getNormalizedPostsQuery } from './get-normalized-posts-query';
+
+/**
+ * Returns a serialized posts query
+ *
+ * @param  {object} query  Posts query
+ * @param  {number} siteId Optional site ID
+ * @returns {string}        Serialized posts query
+ */
+export function getSerializedPostsQuery( query = {}, siteId ) {
+	const normalizedQuery = getNormalizedPostsQuery( query );
+	const serializedQuery = JSON.stringify( normalizedQuery );
+
+	if ( siteId ) {
+		return [ siteId, serializedQuery ].join( ':' );
+	}
+
+	return serializedQuery;
+}
